// src/types/rows/OrcamentoPorLoja.ts
export interface OrcamentoPorLoja {
  anomes: string;
  faturamento_bruto_previsto: number;
  faturamento_bruto_real: number;
  impostos_vendas_previsto: number;
  impostos_previsto_valor: number;
  impostos_vendas_real: number;
  cmv_previsto: number;
  cmv_real: number;
  diferenca_inv_previsto: number;
  diferenca_inv_real: number;
  lucro_bruto_previsto: number;
  lucro_bruto_real: number;
}